SQUASH!!
Rough draft of SMT & continuation support

- Still need to remove feature gating stuff.
