+++ trace_decoder/src/compact/complex_test_payloads.rs
@@ -3,13 +3,13 @@ use mpt_trie::partial_trie::PartialTrie;
 use super::{
     compact_prestate_processing::{
         process_compact_prestate, process_compact_prestate_debug, CompactParsingResult,
-        PartialTriePreImages, ProcessedCompactOutput,
+        MptPartialTriePreImages, ProcessedCompactOutput,
     },
     compact_to_partial_trie::StateTrieExtractionOutput,
 };
 use crate::{
     aliased_crate_types::AccountRlp,
-    trace_protocol::TrieCompact,
+    trace_protocol::MptTrieCompact,
     types::{HashedAccountAddr, TrieRootHash, EMPTY_TRIE_HASH},
     utils::{print_value_and_hash_nodes_of_storage_trie, print_value_and_hash_nodes_of_trie},
 };
@@ -32,7 +32,7 @@ pub(crate) const TEST_PAYLOAD_6: TestProtocolInputAndRoot = TestProtocolInputAnd
     root_str: "135a0c66146c60d7f78049b3a3486aae3e155015db041a4650966e001f9ba301",
 };
 
-type ProcessCompactPrestateFn = fn(TrieCompact) -> CompactParsingResult<ProcessedCompactOutput>;
+type ProcessCompactPrestateFn = fn(MptTrieCompact) -> CompactParsingResult<ProcessedCompactOutput>;
 
 pub(crate) struct TestProtocolInputAndRoot {
     pub(crate) byte_str: &'static str,
@@ -55,7 +55,7 @@ impl TestProtocolInputAndRoot {
         let protocol_bytes = hex::decode(self.byte_str).unwrap();
         let expected_hash = TrieRootHash::from_slice(&hex::decode(self.root_str).unwrap());
 
-        let out = match process_compact_prestate_f(TrieCompact(protocol_bytes)) {
+        let out = match process_compact_prestate_f(MptTrieCompact(protocol_bytes)) {
             Ok(x) => x,
             Err(err) => panic!("{}", err.to_string()),
         };
