+++ trace_decoder/src/decoding_smt.rs
@@ -1,0 +1,9 @@
+use thiserror::Error;
+
+/// Stores the result of parsing tries. Returns a [TraceParsingError] upon
+/// failure.
+pub type SmtTraceParsingResult<T> = Result<T, SmtTraceParsingError>;
+
+/// Error from parsing an SMT trie.
+#[derive(Clone, Debug, Error)]
+pub enum SmtTraceParsingError {}
